backend/requirements.txt
Django==4.2.16
djangorestframework
django-cors-headers
rest_framework_simplejwt
gunicorn
mysqlclient
dj-database-url
whitenoise
psycopg2-binary
